requirements.txt
python-jobspy>=1.1.0
sentence-transformers>=2.2.0
scikit-learn>=1.0
umap-learn>=0.5.0
pandas>=1.3.0
numpy>=1.21.0
plotly>=5.0.0
folium>=0.14.0
matplotlib>=3.5.0
seaborn>=0.12.0
fpdf2>=2.7.0
geopy>=2.3.0
pyyaml>=6.0
schedule>=1.2.0
wordcloud>=1.9.0
kaleido>=0.2.0
pytest>=7.0.0
flake8>=6.0.0
